699: First draft of filter which implements validation rules for FAPI DCR
- Adding FAPIAdvancedDCRValidationFilter which applies validation to a DCR request to ensure that only DCR requests which would result in a FAPI complaint OAuth2 client are allowed through
- Installed this filter into the DCR route before the OB specific filters run
- Renaming module: securebanking-openbanking-uk-gateway-jwks to secure-api-gateway-ig-extensions
- Adding SecureApiGatewayClassAliasResolver which enables extension classes in this module to be referenced in IG config by their short class name
- Moved existing classes under base pkg: com.forgerock.sapi.gateway

https://github.com/SecureApiGateway/SecureApiGateway/issues/699
