+++ config/7.1.0/securebanking/ig/config/dev/config/config.json
@@ -373,7 +373,7 @@
     },
     {
       "name": "OBJwkSetService",
-      "type": "com.forgerock.securebanking.uk.gateway.jwks.cache.caffeine.CaffeineCachingJwkSetService",
+      "type": "CaffeineCachingJwkSetService",
       "config": {
         "maxCacheEntries": 500,
         "expireAfterWriteDuration": "24 hours"
@@ -381,7 +381,7 @@
     },
     {
       "name": "RsaJwtSignatureValidator",
-      "type": "com.forgerock.securebanking.uk.gateway.jws.RsaJwtSignatureValidator"
+      "type": "RsaJwtSignatureValidator"
     }
   ],
   "monitor": true
