+++ config/7.1.0/securebanking/ig/config/prod/config/config.json
@@ -200,7 +200,7 @@
     },
     {
       "name": "OBJwkSetService",
-      "type": "com.forgerock.securebanking.uk.gateway.jwks.cache.caffeine.CaffeineCachingJwkSetService",
+      "type": "CaffeineCachingJwkSetService",
       "config": {
         "maxCacheEntries": 500,
         "expireAfterWriteDuration": "1 hour"
@@ -208,7 +208,7 @@
     },
     {
       "name": "RsaJwtSignatureValidator",
-      "type": "com.forgerock.securebanking.uk.gateway.jws.RsaJwtSignatureValidator"
+      "type": "RsaJwtSignatureValidator"
     }
   ],
   "monitor": true
