+++ config/7.1.0/securebanking/ig/routes/routes-service/03-ob-dcr-register-tpp.json
@@ -39,6 +39,14 @@
           }
         },
         {
+          "comment": "Validate that the request is FAPI compliant",
+          "name": "FAPIAdvancedDCRValidationFilter",
+          "type": "FAPIAdvancedDCRValidationFilter",
+          "config": {
+            "certificateHeader": "ssl-client-cert"
+          }
+        },
+        {
           "comment": "Pre process registration request",
           "name": "ProcessRegistration",
           "type": "ScriptableFilter",
