+++ secure-api-gateway-ig-extensions/src/main/java/com/forgerock/sapi/gateway/dcr/Validator.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.forgerock.sapi.gateway.dcr;
+
+/**
+ * Validator to be used to valid objects as part of Dynamic Client Registration
+ * @param <T>
+ */
+@FunctionalInterface
+public interface Validator<T> {
+
+    /**
+     * Method which applies validation rules to an object.
+     *
+     * If validation passes then the method will return as normal, if validation fails then a ValidationException
+     * must be thrown.
+     *
+     * @param t the object to validate
+     * @throws ValidationException must be thrown if validation fails
+     */
+    void validate(T t) throws ValidationException;
+
+}
