+++ secure-api-gateway-ig-extensions/src/main/resources/META-INF/services/org.forgerock.openig.alias.ClassAliasResolver
@@ -1,0 +1,9 @@
+#
+# Copyright 2014-2022 ForgeRock AS. All Rights Reserved
+#
+# Use of this code requires a commercial software license with ForgeRock AS.
+# or with one of its affiliates. All use shall be exclusively subject
+# to such license between the licensee and ForgeRock AS.
+#
+
+com.forgerock.sapi.gateway.SecureApiGatewayClassAliasResolver
