--- securebanking-openbanking-uk-gateway-jwks/src/main/java/com/forgerock/securebanking/uk/gateway/jwks/RestJwkSetService.java
@@ -1,58 +1,0 @@
-/*
- * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.securebanking.uk.gateway.jwks;
-
-import java.net.URL;
-
-import org.forgerock.json.jose.exceptions.FailedToLoadJWKException;
-import org.forgerock.json.jose.jwk.JWK;
-import org.forgerock.json.jose.jwk.JWKSet;
-import org.forgerock.json.jose.jwk.JWKSetParser;
-import org.forgerock.util.Function;
-import org.forgerock.util.promise.Promise;
-
-/**
- * JwkSetService which fetches JWKSet data from a REST endpoint, implementation is delegated to {@link JWKSetParser}
- */
-public class RestJwkSetService implements JwkSetService {
-
-    private final JWKSetParser jwkSetParser;
-
-    public RestJwkSetService(JWKSetParser jwkSetParser) {
-        this.jwkSetParser = jwkSetParser;
-    }
-
-    @Override
-    public Promise<JWKSet, FailedToLoadJWKException> getJwkSet(URL jwkStoreUrl) {
-        return jwkSetParser.jwkSetAsync(jwkStoreUrl);
-    }
-
-    @Override
-    public Promise<JWK, FailedToLoadJWKException> getJwk(URL jwkStoreUrl, String keyId) {
-        return getJwkSet(jwkStoreUrl).then(findJwkInJwkSet(keyId));
-    }
-
-    private static Function<JWKSet, JWK, FailedToLoadJWKException> findJwkInJwkSet(String keyId) {
-        return jwkSet -> {
-            final JWK jwk = jwkSet.findJwk(keyId);
-            if (jwk != null) {
-                return jwk;
-            } else {
-                throw new FailedToLoadJWKException("Failed to find keyId: " + keyId + " in JWKSet");
-            }
-        };
-    }
-}
