--- securebanking-openbanking-uk-gateway-jwks/src/test/java/com/forgerock/securebanking/uk/gateway/jwks/cache/HashMapCache.java
@@ -1,42 +1,0 @@
-/*
- * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.securebanking.uk.gateway.jwks.cache;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Cache implementation for test purposes, delegates to a HashMap
- */
-public class HashMapCache<K, V> implements Cache<K, V> {
-
-    private final Map<K, V> cache = new HashMap<>();
-
-    @Override
-    public V get(K key) {
-        return cache.get(key);
-    }
-
-    @Override
-    public void put(K key, V value) {
-        cache.put(key, value);
-    }
-
-    @Override
-    public void invalidate(K key) {
-        cache.remove(key);
-    }
-}
